requirements.txt
Django==5.1.2
whitenoise==6.7.0
django-unfold==0.40.0
django-import-export==4.1.1